refactor: move BE typeguards to its own file mimicking FE structure
